+++ src/boards/board_avr.h
@@ -64,11 +64,11 @@ class board_avr: virtual public board
       void UpdateHardware(void);
       void MStep(void);
       void MStepResume(void);
-      int MTestBP(unsigned short bp);
+      int MTestBP(unsigned int bp);
       void MReset(int flags);
       unsigned short * MGetProcID_p(void);
-      unsigned short  MGetPC(void);
-      void MSetPC(unsigned short pc);
+      unsigned int  MGetPC(void);
+      void MSetPC(unsigned int pc);
       unsigned char * MGetRAM_p(void);
       unsigned char * MGetROM_p(void);
       unsigned char * MGetCONFIG_p(void);
